Rename files to avoid conflicts with iOS App.
