+++ RuleEngineMain/src/ruleengine/ConstantTemplate_sbx.cpp
