+++ RuleEngineMain/src/ruleengine/KonceptInfo_sbx.h
